refactor(consys-proto): organize classes into per-resource subpackages
Move bindings/codecs/schema machinery out of the flat proto package into
subpackages by concern: schema/ (writer, reader, both caches), codec/
(encoder, decoder), and datastreams/ observations/ controlstreams/ commands/
for the resource bindings. Service/format/infra classes (ProtoFormat,
ConSysApiProtoService, *Config, *Descriptor, Activator) stay at the root so
the deployment-config FQN, the IModuleProvider SPI entry, and the OSGi
Bundle-Activator keep resolving.

Add the now-required cross-package imports and make the four binding name
constants (OBS_PACKAGE/OBS_MESSAGE, CMD_PACKAGE/CMD_MESSAGE) public so
ProtoFormat reads them across the boundary. Tests mirror the layout.
No behavior change; full module suite green.
