+++ services/sensorhub-service-consys-proto/src/main/java/org/sensorhub/impl/service/consys/proto/codec/ProtoObsEncoder.java
@@ -14,8 +14,9 @@ Author: Ian Patterson <ian.patterson@georobotix.us>
 
 ******************************* END LICENSE BLOCK ***************************/
 
-package org.sensorhub.impl.service.consys.proto;
+package org.sensorhub.impl.service.consys.proto.codec;
 
+import org.sensorhub.impl.service.consys.proto.schema.ProtoSchemaWriter;
 import java.time.Instant;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.FieldDescriptor;
